Change tab select into devided components, add quest tab panel
