+++ components/FreelancerStatsTabs/TabPanel.tsx
@@ -1,0 +1,26 @@
+import { Box, Typography } from "@mui/material"
+import { ReactNode } from "react"
+
+interface TabPanelProps {
+  children?: ReactNode
+  index: number
+  value: number
+}
+
+export function TabPanel({ children, value, index, ...other }: TabPanelProps) {
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && (
+        <Box sx={{ p: 3 }}>
+          <Typography>{children}</Typography>
+        </Box>
+      )}
+    </div>
+  )
+}
