--- components/Tabs/ExpirienceAndReviews.tsx.tsx
@@ -1,5 +1,0 @@
-import { Box } from "@mui/material"
-
-export function ExperienceAndReviews(): JSX.Element {
-  return <Box>Experience and reviews</Box>
-}
