--- components/Tabs/ProfileStats.tsx
@@ -1,5 +1,0 @@
-import { Box } from "@mui/material"
-
-export function ProfileStats(): JSX.Element {
-  return <Box>Profile Page</Box>
-}
